first round of fixes while running tests
